+++ src/game/managers/VOSwordDraw.ts
@@ -1,5 +1,5 @@
 export interface VOSwordDraw {
 	x: number;
 	y: number;
-	gfxTile: number;
+	tileId: number;
 }